JavaScript: Move extraction test cases into separate file
And remove unittest boilerplate while at it
